requirements.txt
Django>=4.2
requests>=2.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
cerebras-cloud-sdk==1.50.1
pytest-django>=4.5.0